UserKeyPressed (works for multiple blue pin pressing)
i defined "HAL_GPIO_EXTI_Callback" inside main and i used it to set the flag.

then i defined a condition inside the main.c loop to check the flag.
if it gain the value of 1, then the microcontroller will transmit "user key pressed" and it will reset the flag after that.